优化菜单选中判断逻辑
