+++ src/layouts/components/MenuContent.tsx
@@ -1,6 +1,7 @@
 import { defineComponent, PropType, computed, h } from 'vue';
 import { prefix } from '@/config/global';
 import { MenuRoute } from '@/interface';
+import { getActive } from '@/router';
 
 const getMenuList = (list: MenuRoute[], basePath?: string): MenuRoute[] => {
   if (!list) {
@@ -34,36 +35,33 @@ const renderIcon = (item) => {
   return () => '';
 };
 
-const useRenderNav = (list: Array<MenuRoute>) => {
+const getPath = (active, item) => {
+  if (active.startsWith(item.path)) {
+    return active;
+  }
+  return item.meta?.single ? item.redirect : item.path;
+};
+
+const useRenderNav = (active: string, list: Array<MenuRoute>) => {
   return list.map((item) => {
     if (!item.children || !item.children.length || item.meta?.single) {
       const href = item.path.match(/(http|https):\/\/([\w.]+\/?)\S*/);
       if (href) {
         return (
-          <t-menu-item
-            href={href?.[0]}
-            name={item.path}
-            value={item.meta?.single ? item.redirect : item.path}
-            icon={renderIcon(item)}
-          >
+          <t-menu-item href={href?.[0]} name={item.path} value={getPath(active, item)} icon={renderIcon(item)}>
             {item.title}
           </t-menu-item>
         );
       }
       return (
-        <t-menu-item
-          name={item.path}
-          value={item.meta?.single ? item.redirect : item.path}
-          to={item.path}
-          icon={renderIcon(item)}
-        >
+        <t-menu-item name={item.path} value={getPath(active, item)} to={item.path} icon={renderIcon(item)}>
           {item.title}
         </t-menu-item>
       );
     }
     return (
       <t-submenu name={item.path} value={item.path} title={item.title} icon={renderIcon(item)}>
-        {item.children && useRenderNav(item.children)}
+        {item.children && useRenderNav(active, item.children)}
       </t-submenu>
     );
   });
@@ -77,6 +75,7 @@ export default defineComponent({
     },
   },
   setup(props) {
+    const active = computed(() => getActive());
     const list = computed(() => {
       const { navData } = props;
       return getMenuList(navData);
@@ -84,11 +83,12 @@ export default defineComponent({
 
     return {
       prefix,
+      active,
       list,
       useRenderNav,
     };
   },
   render() {
-    return <div>{this.useRenderNav(this.list)}</div>;
+    return <div>{this.useRenderNav(this.active, this.list)}</div>;
   },
 });
